+++ src/app/modules/map/ol/lib/charts/layer-raster-chart.component.ts
@@ -15,7 +15,7 @@ import { XYZ } from 'ol/source';
 import { initPMTilesXYZLayer } from './pmtiles-utils';
 import { osmLayer } from '../util';
 import { MapComponent } from '../map.component';
-import { resolveLayerMaxZoom } from './zoom-utils';
+import { extentFromBounds, resolveLayerMaxZoom } from './chart-utils';
 
 import { FBChart } from 'src/app/types';
 
@@ -102,6 +102,7 @@ export class RasterChartLayerComponent implements OnDestroy {
         }
       }
       if (this.layer) {
+        this.layer.setExtent(extentFromBounds(chart[1].bounds));
         this.layer.set('id', chart[0]);
         this.layer.set('chartId', chart[0]);
         this.layer.set('chartType', chart[1].type);
@@ -123,6 +124,7 @@ export class RasterChartLayerComponent implements OnDestroy {
       this.layer.setMinZoom(minZ);
       this.layer.setMaxZoom(layerMaxZ);
       this.layer.setOpacity(chart[1].defaultOpacity ?? 1);
+      this.layer.setExtent(extentFromBounds(chart[1].bounds));
     }
     map.render();
   }